requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary
haversine